@@ -1,36 +1,31 @@
 package com.bowenchin.android.noter;
 
 import android.app.AlertDialog;
-import android.content.ContentUris;
-import android.content.Context;
 import android.content.DialogInterface;
-import android.content.Intent;
 import android.database.Cursor;
-import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
+import androidx.fragment.app.Fragment;
+import androidx.loader.app.LoaderManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.support.v4.content.CursorLoader;
-import android.support.v4.content.Loader;
+import androidx.loader.content.CursorLoader;
+import androidx.loader.content.Loader;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.bowenchin.android.noter.provider.TaskProvider;
-import com.bowenchin.android.noter.util.ReminderManager;
 import com.github.brnunes.swipeablerecyclerview.SwipeableRecyclerViewTouchListener;
 
 /**
  * A simple {@link Fragment} subclass.
  */
 
-public class TaskListFragment extends Fragment implements android.support.v4.app.LoaderManager.LoaderCallbacks<Cursor> {
+public class TaskListFragment extends Fragment implements LoaderManager.LoaderCallbacks<Cursor> {
     RecyclerView recyclerView;
     TaskListAdapter adapter;
     private TextView emptyView;
